backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
sqlalchemy==2.0.23
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78